requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.7
django-cors-headers==3.5.0
django-rest-swagger==2.2.0
django-stubs==1.7.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djangorestframework-stubs==1.1.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mypy==0.790
mypy-extensions==0.4.3
openapi-codec==1.3.2
Pillow==7.0.0
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
sqlparse==0.3.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.7
